Rename and move classes around
